unregister from broadcast receiver in the onDestroy
Before we tried to unregister from broadcast receiver in the onPause callback,
which means that when we are out of the focus on the MainActivity we loosing
synchronization with the server to fetch the latest weather info. If I put unregistration
from broadcast receiver in the onStop callback I won't get the latest weather data
when I'm interacting with another activity.